Fix examples with new interface
