+++ acados/dense_qp/dense_qp_common.c
@@ -19,21 +19,21 @@
 
 // external
 #include <assert.h>
-// blasfeo
-#include "blasfeo/include/blasfeo_target.h"
-#include "blasfeo/include/blasfeo_common.h"
-#include "blasfeo/include/blasfeo_d_aux.h"
-#include "blasfeo/include/blasfeo_d_aux_ext_dep.h"
 // hpipm
 #include "hpipm/include/hpipm_d_dense_qp.h"
 #include "hpipm/include/hpipm_d_dense_qp_sol.h"
 #include "hpipm/include/hpipm_d_dense_qp_res.h"
 #include "hpipm/include/hpipm_d_dense_qp_ipm.h"
 #include "hpipm/include/hpipm_d_dense_qp_kkt.h"
+// blasfeo
+#include "blasfeo/include/blasfeo_target.h"
+#include "blasfeo/include/blasfeo_common.h"
+#include "blasfeo/include/blasfeo_d_aux.h"
+#include "blasfeo/include/blasfeo_d_aux_ext_dep.h"
+#include "blasfeo/include/blasfeo_d_blas.h"
 // acados
 #include "acados/utils/types.h"
 #include "acados/dense_qp/dense_qp_common.h"
-
 
 
 int dense_qp_dims_calculate_size()
