+++ interfaces/acados_c/legacy_create.c
@@ -23,6 +23,26 @@
 
 
 
+ocp_qp_res *create_ocp_qp_res(ocp_qp_dims *dims)
+{
+    int size = ocp_qp_res_calculate_size(dims);
+    void *ptr = acados_malloc(size, 1);
+    ocp_qp_res *qp_res = assign_ocp_qp_res(dims, ptr);
+    return qp_res;
+}
+
+
+
+ocp_qp_res_ws *create_ocp_qp_res_ws(ocp_qp_dims *dims)
+{
+    int size = ocp_qp_res_ws_calculate_size(dims);
+    void *ptr = acados_malloc(size, 1);
+    ocp_qp_res_ws *res_ws = assign_ocp_qp_res_ws(dims, ptr);
+    return res_ws;
+}
+
+
+
 ocp_qp_full_condensing_args *ocp_qp_full_condensing_create_arguments(ocp_qp_dims *dims)
 {
     int size = ocp_qp_full_condensing_calculate_args_size(dims);
