+++ interfaces/acados_c/legacy_create.h
@@ -33,6 +33,10 @@ extern "C" {
 // acados_c
 #include "acados_c/common.h"
 
+ocp_qp_res *create_ocp_qp_res(ocp_qp_dims *dims);
+
+ocp_qp_res_ws *create_ocp_qp_res_ws(ocp_qp_dims *dims);
+
 ocp_qp_full_condensing_args *ocp_qp_full_condensing_create_arguments(ocp_qp_dims *dims);
 
 ocp_qp_full_condensing_memory *ocp_qp_full_condensing_create_memory(ocp_qp_dims *dims, ocp_qp_full_condensing_args *args);
